Extract file import/export to separate hook
